requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.3
schedule==1.2.2
urllib3==2.2.3
Werkzeug==3.1.3
python-dotenv==1.0.1
setuptools==78.1.1

# Linux only
smbus==1.1.post2
gpiozero
sensirion-gas-index-algorithm==3.2.2
sensirion-gas-index-algorithm==3.2.2
sensirion-i2c-sgp4x==0.1.0
sensirion-i2c-sht==0.4.0
